fix(ollama): move OllamaEmbeddingModels back to Ollama client module
